Added sister styling from gif-converter repository, Update App.js accordingly
