+++ package.json
@@ -1,48 +1,60 @@
 {
-    "name": "morse-code-translator",
-    "version": "0.1.0",
-    "private": true,
-    "homepage": "https://Chizaruu.github.io/morse-code-translator/",
-    "dependencies": {
-        "@testing-library/jest-dom": "^5.16.4",
-        "@testing-library/react": "^13.2.0",
-        "@testing-library/user-event": "^13.5.0",
-        "enzyme": "^3.11.0",
-        "node-sass": "^7.0.1",
-        "prop-types": "^15.8.1",
-        "react": "^18.0.0",
-        "react-dom": "^18.1.0",
-        "react-scripts": "5.0.1",
-        "standardized-audio-context-mock": "^9.5.3",
-        "web-vitals": "^2.1.4"
-    },
-    "scripts": {
-        "start": "react-scripts start",
-        "build": "react-scripts build",
-        "test": "react-scripts test",
-        "eject": "react-scripts eject"
-    },
-    "eslintConfig": {
-        "extends": [
-            "react-app",
-            "react-app/jest"
-        ]
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    },
-    "devDependencies": {
-        "@babel/preset-react": "^7.16.7",
-        "babel-jest": "^28.1.0",
-        "nth-check": ">=2.0.1"
-    }
+  "name": "morse-code-translator",
+  "version": "0.1.0",
+  "private": true,
+  "homepage": "https://Chizaruu.github.io/morse-code-translator/",
+  "dependencies": {
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^13.5.0",
+    "enzyme": "^3.11.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.1.0",
+    "react-scripts": "5.0.1",
+    "sass": "^1.51.0",
+    "standardized-audio-context-mock": "^9.5.3",
+    "web-vitals": "^2.1.4"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@babel/preset-react": "^7.16.7",
+    "babel-jest": "^28.1.0",
+    "nth-check": ">=2.0.1"
+  },
+  "description": "A simple program that converts text to morse code.",
+  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Chizaruu/morse-code-translator.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Chizaruu/morse-code-translator/issues"
+  }
 }
