+++ src/App.module.scss
@@ -1,340 +1,12 @@
-// Base font
-$font-family-styles: (
-    heading: (
-        "Roboto Slab",
-        "Helvetica Neue",
-        Arial,
-        sans-serif,
-    ),
-    base: (
-        "Roboto",
-        "Helvetica Neue",
-        Arial,
-        sans-serif,
-    ),
-);
-
-// Font weights
-$font-weights: (
-    bolder: 900,
-    bold: 700,
-    normal: 500,
-    light: 300,
-);
-
-// Font size
-$font-sizes-headings: (
-    h1: 3rem,
-    h2: 2.5rem,
-    h3: 2rem,
-    h4: 1.75rem,
-    h5: 1.5rem,
-    h6: 1.25rem,
-);
-
-$font-sizes-text: (
-    base: 0.75rem,
-    small: 0.625rem,
-    medium: 0.875rem,
-    large: 1rem,
-    larger: 1.25rem,
-    enlarge: 1.5rem,
-);
-
-// Line-height
-$line-heights: (
-    loose: 1.75,
-    normal: 1.5,
-    tight: 1.25,
-);
-
-// Letter-spacing
-$letter-spacings: (
-    extra-wide: 0.1rem,
-    wide: 0.05rem,
-    normal: 0,
-    tight: -0.05rem,
-);
-
-%font-headings {
-    font-family: map-get($font-family-styles, heading);
-    line-height: map-get($line-heights, normal);
-    letter-spacing: map-get($letter-spacings, normal);
-    font-weight: map-get($font-weights, bold);
-    white-space: nowrap;
-}
-
-$light-theme-colors: (
-    primary: #ffffff,
-    secondary: #f3f4f6,
-    tertiary: #00ba34,
-    quaternary: #b0dabc,
-    quinary: #05080e67,
-);
-
-$dark-theme-colors: (
-    primary: #000000,
-    secondary: #222222,
-    tertiary: #17cb49,
-    quaternary: #6acf86,
-    quinary: #f3f4f667,
-);
-
-//Groups of font colors
-$light-theme-font-colors: (
-    primary: black,
-    secondary: #00ba34,
-);
-
-$dark-theme-font-colors: (
-    primary: #ffffff,
-    secondary: #17cb49,
-);
-
-//Group of special colors
-$special-colors: (
-    heart: #e74c3c,
-);
-
-$device-sizes: (
-    phone: 450px,
-    tablet: 880px,
-    desktop: 1024px,
-);
-
-@mixin phone {
-    @media only screen and (max-width: map-get($device-sizes, phone)) {
-        @content;
-    }
-}
-
-@mixin tablet {
-    @media only screen and (min-width: (map-get($device-sizes, phone) +1px)) and (max-width: map-get($device-sizes, tablet)) {
-        @content;
-    }
-}
-
-@mixin desktop {
-    @media only screen and (min-width: (map-get($device-sizes, tablet) +1px)) {
-        @content;
-    }
-}
-
-//Themes
-@mixin light {
-    @media (prefers-color-scheme: light) {
-        @at-root #{selector-nest(':root:not([data-theme=light]):not([data-theme=dark])', &)} {
-            @content;
-        }
-    }
-}
-
-@mixin dark {
-    @media (prefers-color-scheme: dark) {
-        @at-root #{selector-nest(':root:not([data-theme=light]):not([data-theme=dark])', &)} {
-            @content;
-        }
-    }
-}
-
-//Technicals
-@mixin display-image($device-size) {
-    @if ($device-size !=map-get($device-sizes, phone)) {
-        @if ($device-size !=map-get($device-sizes, tablet)) {
-            width: 50%;
-            height: 100%;
-        } @else {
-            width: 75%;
-            height: 100%;
-        }
-    } @else {
-        width: 0%;
-        height: 0%;
-    }
-}
-
-//Data-Theme
-%data-theme {
-    @include light {
-        color: map-get($light-theme-font-colors, primary);
-    }
-
-    @include dark {
-        color: map-get($dark-theme-font-colors, primary);
-    }
-}
+@use "./scss/typography.module";
+@use "./scss/background.module";
+@use "./scss/mixins.module";
+@use "./scss/button.module";
 
 .App {
+    scroll-behavior: smooth;
     margin: 0;
     padding: 0;
-
-    background-color: #222222;
-    color: #ffffff;
-
-    button {
-        text-align: center;
-        padding: 5px;
-        border-radius: 5px;
-        font-size: 1.5rem;
-        width: 10vw;
-        margin: 5px;
-        height: 8vh;
-
-        &:hover {
-            cursor: pointer;
-        }
-
-        @include light {
-            background-color: map-get($light-theme-colors, tertiary);
-        }
-
-        @include dark {
-            background-color: map-get($dark-theme-colors, tertiary);
-        }
-    }
-
-    //headings
-    h1,
-    h2,
-    h3,
-    h4,
-    h5,
-    h6 {
-        @extend %font-headings;
-        @extend %data-theme;
-    }
-
-    h1 {
-        font-size: map-get($font-sizes-headings, h1);
-
-        @include phone {
-            font-size: map-get($font-sizes-headings, h4);
-        }
-    }
-
-    h2 {
-        font-size: map-get($font-sizes-headings, h2);
-
-        @include phone {
-            font-size: map-get($font-sizes-headings, h5);
-        }
-    }
-
-    //text
-    p,
-    a,
-    li,
-    footer {
-        font-family: map-get($font-family-styles, base);
-        line-height: map-get($line-heights, normal);
-
-        @include phone {
-            font-size: map-get($font-sizes-text, small);
-        }
-
-        @include tablet {
-            font-size: map-get($font-sizes-text, medium);
-        }
-
-        @include desktop {
-            font-size: map-get($font-sizes-text, larger);
-        }
-
-        @extend %data-theme;
-    }
-
-    a,
-    footer,
-    button {
-        font-weight: map-get($font-weights, bold);
-    }
-
-    header {
-        display: flex;
-        justify-content: center;
-
-        @include light {
-            background-color: map-get($light-theme-colors, secondary);
-            box-shadow: 0px 0px 7px 0px map-get($dark-theme-colors, secondary);
-        }
-
-        @include dark {
-            background-color: map-get($dark-theme-colors, secondary);
-            box-shadow: 0px 0px 7px 0px map-get($light-theme-colors, secondary);
-        }
-    }
-
-    main {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        text-align: center;
-        padding: 5px;
-        border-radius: 5px;
-        height: 75vh;
-        margin-top: 5vh;
-
-        @include light {
-            color: map-get($light-theme-font-colors, primary);
-        }
-
-        @include dark {
-            color: map-get($dark-theme-font-colors, primary);
-        }
-
-        textarea {
-            width: 15vw;
-            height: 15vh;
-            border: 1px solid #ccc;
-            border-radius: 5px;
-            padding: 5px;
-            resize: none;
-        }
-
-        div {
-            display: flex;
-            flex-direction: row;
-            justify-content: center;
-            align-items: center;
-            width: 100%;
-            height: 10vh;
-        }
-
-        label {
-            font-size: 1.5rem;
-        }
-    }
-
-    footer {
-        position: fixed;
-        padding: 15px;
-        text-align: center;
-        bottom: 0;
-        width: 100%;
-
-        span {
-            display: inline-flex;
-            color: map-get($special-colors, heart);
-            animation: love 0.7s infinite ease-out;
-
-            @keyframes love {
-                0%,
-                100% {
-                    transform: scale(1.3);
-                }
-
-                30% {
-                    transform: scale(1.5);
-                }
-            }
-        }
-
-        @include light {
-            background-color: map-get($light-theme-colors, secondary);
-        }
-
-        @include dark {
-            background-color: map-get($dark-theme-colors, secondary);
-        }
-    }
+    min-height: calc(100% + constant(safe-area-inset-top));
+    min-height: calc(100% + env(safe-area-inset-top));
 }
