+++ src/containers/Footer/Footer.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import styles from "./Footer.module.scss";
+
+export function Footer() {
+    return (
+        <footer className={styles}>
+            Made with <span title="Latvia!~"> ❤️ </span> by Abdul-Kadir Coskun
+        </footer>
+    );
+}
