+++ src/containers/Footer/Footer.module.scss
@@ -1,0 +1,35 @@
+@use "./../../scss/palette.module" as p;
+@use "./../../scss/mixins.module" as m;
+
+footer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    height: 3em;
+
+    span {
+        display: inline-flex;
+        color: map-get(p.$special-colors, heart);
+        animation: love 0.7s infinite ease-out;
+
+        @keyframes love {
+            0%,
+            100% {
+                transform: scale(1.1);
+            }
+
+            30% {
+                transform: scale(1.3);
+            }
+        }
+    }
+
+    @include m.light {
+        background-color: map-get(p.$light-theme-colors, secondary);
+    }
+
+    @include m.dark {
+        background-color: map-get(p.$dark-theme-colors, secondary);
+    }
+}
