+++ src/containers/Header/Header.jsx
@@ -1,0 +1,10 @@
+import React from "react";
+import styles from "./Header.module.scss";
+
+export function Header() {
+    return (
+        <header className={styles}>
+            <h1>Morse Code Translator</h1>
+        </header>
+    );
+}
