+++ src/containers/Main/Main.module.scss
@@ -1,0 +1,59 @@
+@use "./../../scss/palette.module" as p;
+@use "./../../scss/mixins.module" as m;
+
+main {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 43em;
+
+    video {
+        width: 50%;
+        height: 50%;
+    }
+
+    button {
+        margin-top: 10px;
+    }
+
+    span {
+        display: flex;
+        justify-content: space-evenly;
+        align-items: center;
+        width: 10em;
+        margin-bottom: 10px;
+
+        label {
+            width: 5em;
+            height: 100%;
+            text-align: right;
+        }
+
+        input,
+        select {
+            input {
+                color: transparent;
+            }
+
+            width: 65px;
+        }
+    }
+
+    div {
+        display: flex;
+        flex-direction: row;
+        justify-content: space-evenly;
+        align-items: center;
+        width: 60%;
+        height: 80%;
+        div {
+            display: flex;
+            flex-direction: column;
+            justify-content: flex-start;
+            align-items: center;
+            width: 100%;
+            height: 100%;
+        }
+    }
+}
