+++ src/containers/Nav/Nav.module.scss
@@ -1,0 +1,233 @@
+@use "./../../scss/palette.module" as p;
+@use "./../../scss/mixins.module" as m;
+
+@mixin navbar_list_hamburger {
+    position: fixed;
+    top: 0;
+    margin: 4.5vh 0 0 0;
+    padding: 0;
+    flex-direction: column;
+    width: 100%;
+    justify-content: center;
+    align-items: center;
+    display: none;
+
+    & > li {
+        display: flex;
+        justify-content: center;
+        margin: 0;
+        padding: 0.5em 0;
+        width: 100%;
+
+        @include m.light {
+            background-color: map-get(p.$light-theme-colors, secondary);
+        }
+
+        @include m.dark {
+            background-color: map-get(p.$dark-theme-colors, secondary);
+        }
+    }
+
+    & > li:not(:last-child) {
+        @include m.light {
+            border-bottom: 1px solid map-get(p.$light-theme-colors, secondary);
+        }
+
+        @include m.dark {
+            border-bottom: 1px solid map-get(p.$dark-theme-colors, secondary);
+        }
+    }
+}
+
+@mixin navbar_list {
+    &__list {
+        list-style: none;
+
+        @include m.phone {
+            @include navbar_list_hamburger;
+        }
+
+        @include m.tablet {
+            @include navbar_list_hamburger;
+        }
+
+        @include m.desktop {
+            display: flex;
+            justify-content: space-evenly;
+            width: 35vw;
+            padding-right: 5vw;
+            align-items: center;
+            height: 100%;
+        }
+
+        &__item {
+            height: 100%;
+            display: flex;
+            align-items: center;
+
+            @include m.desktop {
+                border-radius: 5px;
+            }
+
+            a {
+                display: flex;
+                justify-content: center;
+                align-items: center;
+                width: 100%;
+                height: 100%;
+            }
+
+            &:hover {
+                @include m.light {
+                    background-color: map-get(p.$light-theme-colors, tertiary);
+                }
+
+                @include m.dark {
+                    background-color: map-get(p.$dark-theme-colors, tertiary);
+                }
+            }
+        }
+    }
+}
+
+@mixin navbar_hamburger {
+    &__hamburger-container {
+        display: flex;
+        height: 100%;
+        width: 30px;
+        cursor: pointer;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+
+        padding-right: 5px;
+
+        &__button,
+        &__button::before,
+        &__button::after {
+            display: block;
+            position: absolute;
+            height: 4px;
+            width: 30px;
+            transition: transform 400ms cubic-bezier(0.23, 1, 0.32, 1);
+            border-radius: 2px;
+
+            @include m.light {
+                background-color: map-get(p.$dark-theme-colors, secondary);
+            }
+
+            @include m.dark {
+                background-color: map-get(p.$light-theme-colors, secondary);
+            }
+        }
+
+        &__button::before {
+            content: "";
+            margin-top: -8px;
+        }
+
+        &__button::after {
+            content: "";
+            margin-top: 8px;
+        }
+    }
+}
+
+.navbar {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    position: fixed;
+    width: 100%;
+    height: 5%;
+    z-index: 1;
+
+    a {
+        text-decoration: none;
+        white-space: nowrap;
+        padding: 0 1vw;
+    }
+
+    @include m.phone {
+        @include navbar-list;
+        @include navbar-hamburger;
+    }
+
+    @include m.tablet {
+        @include navbar-list;
+        @include navbar-hamburger;
+    }
+
+    @include m.desktop {
+        @include navbar-list;
+    }
+
+    @include m.light {
+        background-color: map-get(p.$light-theme-colors, secondary);
+        box-shadow: 0px 0px 7px 0px map-get(p.$dark-theme-colors, secondary);
+    }
+
+    @include m.dark {
+        background-color: map-get(p.$dark-theme-colors, secondary);
+        box-shadow: 0px 0px 7px 0px map-get(p.$light-theme-colors, secondary);
+    }
+}
+
+#hamburger_toggle {
+    display: none;
+}
+
+#hamburger_toggle:checked
+    + .navbar__hamburger_container
+    .navbar__hamburger_container__button::before {
+    margin-top: 0px;
+    transform: rotate(405deg);
+}
+
+#hamburger_toggle:checked
+    + .navbar__hamburger_container
+    .navbar__hamburger_container__button {
+    background: rgba(255, 255, 255, 0);
+}
+
+#hamburger_toggle:checked
+    + .navbar__hamburger_container
+    .navbar__hamburger_container__button::after {
+    margin-top: 0px;
+    transform: rotate(-405deg);
+}
+
+@mixin hamburger_small_display_toggle {
+    #hamburger_toggle:checked ~ .navbar__list {
+        display: flex;
+    }
+
+    #hamburger_toggle ~ .navbar__list li {
+        height: 0;
+        margin: 0;
+        padding: 0;
+        border: 0;
+        transition: height 400ms cubic-bezier(0.23, 1, 0.32, 1);
+    }
+
+    #hamburger_toggle:checked ~ .navbar__list li {
+        @include m.light {
+            border: 1px solid map-get(p.$light-theme-colors, quinary);
+        }
+
+        @include m.dark {
+            border: 1px solid map-get(p.$dark-theme-colors, quinary);
+        }
+
+        height: 3.5em;
+        transition: height 400ms cubic-bezier(0.23, 1, 0.32, 1);
+    }
+}
+
+@include m.phone {
+    @include hamburger_small_display_toggle;
+}
+
+@include m.tablet {
+    @include hamburger_small_display_toggle;
+}
